[#62] petsitter_reservation_list  에서 adapter 패키지 생성하여 파일 나눔. (정리)
