+++ MungNolZa/app/src/main/res/layout/fragment_petsitter_reservation_list_last.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".ui.fragment.PetsitterReservationListLastFragment">
+    tools:context=".ui.petsitter_reservation_list.fragment.PetsitterReservationListLastFragment">
 
     <TextView
         android:id="@+id/textViewPetsitterReservationListLastCount"
